got rid of the unused fields warnings
implemented the TreeSearchCost trait that calls the logl() function. so now it is considered to be used. why the cost() of the trait is considered used is not clear to me, maybe simply bc the function cost() is declared in a different crate and rust stops analyzing beyond that point