Feature/random provider (#88)
* Initial version of the global random

Initial version of the global rng for the whole crate, currently using StdRng
which is cryptographically secure but not necessarily reproducible because it depends
on the system configuration and might change with library versions.

* Generalisation to work with different RNG backends

Generalised the global random implementation to allow different RNG backends.
The crate-wide RNG still uses StdRng.

* Examples for phylo::random use

* Add trait for RandomSource

* Inject rng into topo optimiser for testing

Injecting a rng provider `RandomSource` into topology optimiser to
allow using a fake deterministic rng in tests and remove the `deterministic`
feature.

* FakeGenerator implementation for testing

FakeGenerator is a deterministic struct that implements `RandomSource` for testing

* Inject rng provider into PhyloInfoBuilder

Inject rng provider into PhyloInfoBuilder to pass to the NJ tree builder

* Injecting rng provider into tree building

* RNG usage examples wo global static random

* Add `ntimestamp` to seed rng

* Simplifying FakeGenerator

* FakeGenerator tests

* Add more types to FakeGenerator::gen()

* Fix FakeGenerator::get_probability()

* Fix test names and small refactor

* Remove redundant test

* Remove gen_range from RandomSource

Remove gen_range from RandomSource trait but keep it as an impl for RandomGenerator.

* Fix fake rng test

* Refactor test names

* Remove unnecessary `deterministic` feature

* Use fake rng in benches for determinism

* Remove `deterministic` feature from CI

* Immediate panic for unsupported types in fakerng

* Add tests for FakeGenerator coverage

Added a test that requests an unimplemented type and a gen_bool(p) test.

* RandomGenerator test cleanup

* Add comment to diff seed test about no guarantee

Added a comment to the test with different seeds producing different first
values, since it is not guaranteed but extremely unlikely to fail.

* Add test for generating basic types

Added a test for generating different basic types with RandomSource just to
make sure it does not panic.

* Add .seed() to RandomSource and impl

Add a .seed() method to RandomSource trait and implement it for RandomGenerator and
FakeGenerator, add tests for getting the seed.

* Add char to FakeGenerator, change unsupported test

Added char to FakeGenerator supported types, changed test to crash on isize

* Clean up test code -- remove unnecessary typing

* Add test for DefaultGenerator::default()

Added test for DefaultGenerator::default() which should be initialised
with the current timestamp as the seed.

* Remove unnecessary mut from RandomSource::shuffle

* Rename run_w_rng back to run in topo_optimiser

* Fix topo_optimiser doctest

* Fix tests for CI/CD

* Add comment on current rng choice

* Add new_with_pred tests to topo_optimiser

* Remove tests from coverage

* Add char generation to fake_generator